Add support for boot/disc storage on SD card
Uses the no-OS-FatFS-SD-SPI-RPi-Pico library, and on startup opens
the fs to search for 'umac0.img' or 'umac0ro.img'.  If the former
is found, it's used as a RW image; the latter is RO.  If no images
are found, the code falls back to an in-flash disc image; you can
set up a default that's used if no SD card is inserted, for example.

The wiring defaults to GPIOs 2/3/4/5 for SCK/TX/RX/CS respectively,
which matches the upper-left SPI0 on the Pico board.  The wiring
can be overridden by defining SD_TX, SD_RX, SD_SCK, SD_CS on the
`cmake` build commandline.

The SPI speed defaults to a very conservative 5MHz (at least one of
my cards will only cope with this).  Again, this can be overridden
by defining SD_MHZ on the commandline (to an integer in units of MHz).

Note this also disables the RTC portion of FatFS when building
for RP2350 (which doesn't have an RTC, or APIs for it).