video: Allow GPIO pins for video to be configured
Adds a new build option, -DVIDEO_PIN=<n>, which sets the base GPIO
of Video (followed by syncs).

The video pins (for PIO reasons) need to be contiguous, numbered upward
from the DATA pin.  This has caught a couple of people out, so to make
it easier just define them this way.  That then gives a single
configuration option, the first GPIO, setting it all up.